Simplify package installation services
