+++ src/services.rs
@@ -1,7 +1,6 @@
 pub mod application_command_service;
 pub mod command_validator;
-pub mod multi_package_installation_service;
-pub mod package_installation_service;
+pub mod package_installer;
 pub mod package_list_service;
 pub mod package_validator;
 pub mod validation_command;
